Fix compilation on macOS.

+++ libOTe/TwoChooseOne/SoftSpokenOT/DotSemiHonest.h
@@ -159,7 +159,7 @@ protected:
 
 	static const size_t commSize = commStepSize * superBlkSize; // picked to match the other OTs.
 	size_t chunkSize() const { return 128; }
-	size_t paddingSize() const { return std::max(divCeil(wPadded(), 2), chunkSize()) - chunkSize(); }
+	size_t paddingSize() const { return std::max(size_t(divCeil(wPadded(), 2)), chunkSize()) - chunkSize(); }
 
 	void recvBuffer(Channel& chl, size_t batchSize) { vole->recv(chl, 0, batchSize); }
 	TRY_FORCEINLINE void processChunk(
